new changes

+++ models.py
@@ -15,7 +15,7 @@ def get_now_jalali_with_time():
 class Item(db.Model):
     __tablename__ = 'items'
 
-    id = db.Column(db.Integer,primary_key=True)
+    id = db.Column(db.BigInteger,primary_key=True)
     project_code = db.Column(db.String(80))
     warehouse_location = db.Column(db.String(80))
     row = db.Column(db.String(80))
@@ -39,7 +39,7 @@ class Item(db.Model):
 class ItemHistory(db.Model):
     __tablename__ = 'items_history'
 
-    id = db.Column(db.Integer,primary_key=True)
+    id = db.Column(db.BigInteger,primary_key=True)
     project_code = db.Column(db.String(80))
     warehouse_location = db.Column(db.String(80))
     row = db.Column(db.String(80))
@@ -65,7 +65,7 @@ class ItemHistory(db.Model):
 class ItemLog(db.Model):
     __tablename__ = 'items_log'
 
-    id = db.Column(db.Integer,primary_key=True)
+    id = db.Column(db.BigInteger,primary_key=True)
     project_code = db.Column(db.String(80))
     warehouse_location = db.Column(db.String(80))
     row = db.Column(db.String(80))
@@ -90,7 +90,7 @@ class ItemLog(db.Model):
 class User(UserMixin, db.Model):
     __tablename__ = 'users'
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     username = db.Column(db.String(100), unique=True, nullable=False)
     password = db.Column(db.String(200), nullable=False)
     is_admin = db.Column(db.Boolean, default=False)
@@ -100,7 +100,7 @@ class User(UserMixin, db.Model):
 class Personnel(db.Model):
     __tablename__ = 'personnels'
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     username = db.Column(db.String(100), nullable=False)
     personnel_code = db.Column(db.String(100), nullable=False)
     company = db.Column(db.String(100), nullable=False)
@@ -112,7 +112,7 @@ class Personnel(db.Model):
 class Repair(db.Model):
     __tablename__ = 'repairs'
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     device_type = db.Column(db.String(100), nullable=False)
     model = db.Column(db.String(100), nullable=False)
     serial_number = db.Column(db.String(100), nullable=False)
@@ -125,7 +125,7 @@ class Repair(db.Model):
 class Phone(db.Model):
     __tablename__ = 'phones'
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     username = db.Column(db.String(100), nullable=False)
     place = db.Column(db.String(100), nullable=False)
     phone_number = db.Column(db.String(100), nullable=False)
